merge: develop -> feat/#70/template-provided
